Deleted meta for ignored Samples folder
Should fix this warning:
A meta data file (.meta) exists but its folder 'Packages/com.fastscriptreload/Samples~' can't be found, and has been created. Empty directories cannot be stored in version control, so it's assumed that the meta data file is for an empty directory in version control. When moving or deleting folders outside of Unity, please ensure that the corresponding .meta file is moved or deleted along with it.
UnityEditor.CodeEditorProjectSync:SyncAndOpenSolution ()